backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
stripe
httpx
sqlmodel
psycopg2-binary
google-genai
openai
